vdev Geocodex.Provider behaviour

A behaviour for geocoder service providers.

You can implement the behaviour by using it in a module:

defmodule MyApp.Provider do
  use Geocodex.Provider
end

The following options are available to the use line:

:base_url - The base url of the geocoding service

:default_params - The default params that should be added to every request.

:default_headers - The default headers that should be added to every request.

Here's an example from Geocodex.Provider.USCB:

  use Geocodex.Provider,
    base_url: "https://geocoding.geo.census.gov/geocoder/locations/onelineaddress",
    default_headers: [{"Accept", "application/json"}],
    default_params: [benchmark: "Public_AR_Current"]

Types

address()

address() :: Geocodex.address()

coordinates()

coordinates() :: Geocodex.coordinates()

Callbacks

coordinates(address)

coordinates(address()) :: {:ok, coordinates()}

Invoked to geocode an address into x/y coordinates