vdev PYREx.Accounts.User

API user. Users must register to receive a key for API access.

Data fields:

:email - String. Must be unique. Validations will fail if the record already exists in the database.

:name - String.

:organization - String.

:website - String.

:intended_usage - String. Longer form explanation of how the API will be used.

:is_authorized - Boolean. Is the user associated with the email address authorized to use the API? Authorization may be suspended for misuse.

Functions

authorized?(user)

Returns true if a user is authorized, false if they are not.

Examples

iex> authorized?(%User{is_authorized: true})
true

iex> authorized?(%User{is_authorized: false})
false