vdev PYREx.Districts

The Districts context.

Functions

change_shape(shape)

Returns an %Ecto.Changeset{} for tracking shape changes.

Examples

iex> change_shape(shape)
%Ecto.Changeset{source: %Shape{}}

create_shape(attrs \\ %{})

Creates a shape.

Examples

iex> create_shape(%{field: value})
{:ok, %Shape{}}

iex> create_shape(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_shape(shape)

Deletes a Shape.

Examples

iex> delete_shape(shape)
{:ok, %Shape{}}

iex> delete_shape(shape)
{:error, %Ecto.Changeset{}}

get_shape!(id)

Gets a single shape.

Raises Ecto.NoResultsError if the Shape does not exist.

Examples

iex> get_shape!(123)
%Shape{}

iex> get_shape!(456)
** (Ecto.NoResultsError)

intersecting_shapes(coordinates)

Returns a list of shapes that intersect the given geometry.

Input can be a tuple containing two floats (lat and lon) or a Geo struct.

list_shapes()

Returns the list of shapes.

Examples

iex> list_shapes()
[%Shape{}, ...]

update_shape(shape, attrs)

Updates a shape.

Examples

iex> update_shape(shape, %{field: new_value})
{:ok, %Shape{}}

iex> update_shape(shape, %{field: bad_value})
{:error, %Ecto.Changeset{}}