vdev PYREx.Shapefile

Read shapefile data.

Functions

dbf(shapefile)

exshape_to_geo(polygon)

Transforms an Exshape.Shp.Polygon.t() to a Geo.Polygon.t().

Examples

iex> PYREx.Shapefile.exshape_to_geo(%Exshape.Shp.Polygon{points: [[[
...>   %Exshape.Shp.Point{x: -96.639704, y: 42.737071},
...>   %Exshape.Shp.Point{x: -96.635886, y: 42.741002},
...>   %Exshape.Shp.Point{x: -96.632314, y: 42.745641}
...> ]]]})
%Geo.MultiPolygon{coordinates: [[[
  {42.737071, -96.639704},
  {42.741002, -96.635886},
  {42.745641, -96.632314}
]]], srid: 3857}

from_zip(shapefile)

Decodes data from a Zip archived shapefile in the current working directory.

map_download(filename, base_url, opts \\ [])

Downloads the given shapefile from the U.S. Census Bureau and maps the data for database insertion. The file is not saved to the file system.

map_shapes(shapes)

Parses unzipped shapefile data into maps.

shp(shapefile)

srid()

The SRID projection used by the application. Returns the integer value 3857.

Examples

iex> PYREx.Shapefile.srid()
3857