vdev PYREx.Sources

The sources for the PYREx database.

Sources are fetched from a remote git repository and require an internet connection to return results.

Functions

get(resource)

Get a list of sources from the remote repo.

shapefiles()

The url paths for shapefile data.