vdev PYREx.TestFixtures behaviour

A module for defining fixtures that can be used in tests. This module can be used with a list of fixtures to apply as parameters:

use PYREx.TestFixtures, [:user]

Callbacks

user_fixture(attrs) (optional)

user_fixture(attrs :: map()) :: PYREx.Accounts.User.t()

Returns a valid user after inserting it into the database.

A map of attrs can be passed as an argument to override the defaults.

Examples

iex>user_fixture()
%PYREx.Accounts.User{}

iex>user_fixture(%{is_authorized: false})
%PYREx.Accounts.User{is_authorized: false}