vdev PYRExWeb.Router.Helpers

Module with named helpers generated from PYRExWeb.Router.

Functions

accounts_user_path(conn_or_endpoint, action)

accounts_user_path(conn_or_endpoint, action, params)

accounts_user_path(conn_or_endpoint, action, id, params)

accounts_user_url(conn_or_endpoint, action)

accounts_user_url(conn_or_endpoint, action, params)

accounts_user_url(conn_or_endpoint, action, id, params)

home_path(conn_or_endpoint, action)

home_path(conn_or_endpoint, action, params)

home_url(conn_or_endpoint, action)

home_url(conn_or_endpoint, action, params)

jurisdiction_path(conn_or_endpoint, action)

jurisdiction_path(conn_or_endpoint, action, params)

jurisdiction_path(conn_or_endpoint, action, id, params)

jurisdiction_url(conn_or_endpoint, action)

jurisdiction_url(conn_or_endpoint, action, params)

jurisdiction_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.