vdev PYREx.Geographies

The Geographies context.

Functions

change_jurisdiction(jurisdiction)

Returns an %Ecto.Changeset{} for tracking jurisdiction changes.

Examples

iex> change_jurisdiction(jurisdiction)
%Ecto.Changeset{source: %Jurisdiction{}}

change_shape(shape)

Returns an %Ecto.Changeset{} for tracking shape changes.

Examples

iex> change_shape(shape)
%Ecto.Changeset{source: %Shape{}}

create_jurisdiction(attrs \\ %{})

Creates a jurisdiction.

Examples

iex> create_jurisdiction(%{field: value})
{:ok, %Jurisdiction{}}

iex> create_jurisdiction(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_shape(attrs \\ %{})

Creates a shape.

Examples

iex> create_shape(%{field: value})
{:ok, %Shape{}}

iex> create_shape(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_jurisdiction(jurisdiction)

Deletes a Jurisdiction.

Examples

iex> delete_jurisdiction(jurisdiction)
{:ok, %Jurisdiction{}}

iex> delete_jurisdiction(jurisdiction)
{:error, %Ecto.Changeset{}}

delete_shape(shape)

Deletes a Shape.

Examples

iex> delete_shape(shape)
{:ok, %Shape{}}

iex> delete_shape(shape)
{:error, %Ecto.Changeset{}}

get_jurisdiction!(id)

Gets a single jurisdiction.

Raises Ecto.NoResultsError if the Jurisdiction does not exist.

Examples

iex> get_jurisdiction!(123)
%Jurisdiction{}

iex> get_jurisdiction!(456)
** (Ecto.NoResultsError)

get_shape!(id)

Gets a single shape.

Raises Ecto.NoResultsError if the Shape does not exist.

Examples

iex> get_shape!(123)
%Shape{}

iex> get_shape!(456)
** (Ecto.NoResultsError)

intersecting_jurisdictions(address)

Returns a list of Jursidictions whose shapes intersect the given geometry.

Input can be a tuple containing two strings or floats (lat and lon), a binary street address, or a Geo struct.

Examples

intersecting_jurisdictions({33.184123, -88.317135})
#=> [%PYREx.Geographies.Jurisdiction{
      __meta__: #Ecto.Schema.Metadata<:loaded, "jurisdictions">,
      geoid: "01",
      id: "pyr-jurisdiction/type:us_state/country:us/pyrgeoid:G400001/statefp:01",
      inserted_at: ~N[2019-03-24 00:23:52],
      name: "Alabama",
      shape: #Ecto.Association.NotLoaded<association :shape is not loaded>,
      statefp: "01",
      mtfcc: G4000,
      pyrgeoid: G400001,
      type: "us_state",
      updated_at: ~N[2019-03-24 00:23:52]
    }, %PYREx.Geographies.Jurisdiction{...}, ...]

intersecting_shapes(address)

Returns a list of shapes that intersect the given geometry.

Input can be a tuple containing two strings or floats (lat and lon), a binary street address, or a Geo struct.

list_jurisdictions()

Returns the list of jurisdictions.

Examples

iex> list_jurisdictions()
[%Jurisdiction{}, ...]

list_shapes()

Returns the list of shapes.

Examples

iex> list_shapes()
[%Shape{}, ...]

update_jurisdiction(jurisdiction, attrs)

Updates a jurisdiction.

Examples

iex> update_jurisdiction(jurisdiction, %{field: new_value})
{:ok, %Jurisdiction{}}

iex> update_jurisdiction(jurisdiction, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_shape(shape, attrs)

Updates a shape.

Examples

iex> update_shape(shape, %{field: new_value})
{:ok, %Shape{}}

iex> update_shape(shape, %{field: bad_value})
{:error, %Ecto.Changeset{}}